First commit to landscape.yml
Please review this first commit to see if it's in line with what is expected. This project did not have an SVG image of its logo so I created one with an online converter that actually hosts the new SVG file. Please let me know otherwise.

I've left the Crunchbase link section empty because Augur Labs doesn't have a Crunchbase profile. 

Signed-off-by: Jordi Mon Companys <MordodeMaru@users.noreply.github.com>